Revert "Move clinic profile editing to dedicated route"
This reverts commit d524582d7a37809f4cc3005d44b7f60f8dcbc5a2.
