+++ app/routes.js
@@ -8,7 +8,6 @@ import AppComponent from './pages/app';
 import BrowserWarning from './pages/browserwarning';
 import ClinicDetails from './pages/clinicdetails';
 import ClinicAdmin from './pages/clinicadmin';
-import ClinicProfile from './pages/clinicprofile';
 import ClinicWorkspace from './pages/clinicworkspace';
 import { TideDashboard } from './pages/dashboard';
 import ClinicInvite from './pages/clinicinvite';
@@ -134,7 +133,6 @@ export const requireAuth = (api, cb = _.noop) => (dispatch, getState) => {
           const restrictedClinicUIRoutes = [
             '/clinic-admin',
             '/clinic-invite',
-            '/clinic-profile',
             '/clinic-workspace',
             '/clinician-edit',
             '/prescriptions',
@@ -192,7 +190,6 @@ export const requireAuth = (api, cb = _.noop) => (dispatch, getState) => {
             '/clinic-admin',
             '/clinic-details',
             '/clinic-invite',
-            '/clinic-profile',
             '/clinic-workspace',
             '/clinician-edit',
             '/prescriptions',
@@ -400,7 +397,6 @@ export const getRoutes = (appContext) => {
           <Route path='/terms' render={routeProps => (<Terms {...routeProps} {...props} />)} />
           <Route path='/signup' render={routeProps => (<Gate onEnter={boundRequireNoAuth} key={routeProps.match.path}><Signup {...routeProps} {...props} /></Gate>)} />
           <Route path='/clinic-admin' render={routeProps => (<Gate onEnter={boundRequireAuth} key={routeProps.match.path}><ClinicAdmin {...routeProps} {...props} /></Gate>)} />
-          <Route path='/clinic-profile' render={routeProps => (<Gate onEnter={boundRequireAuth} key={routeProps.match.path}><ClinicProfile {...routeProps} {...props} /></Gate>)} />
           <Route path='/clinic-details/:action' render={routeProps => (<Gate onEnter={boundRequireAuth} key={routeProps.match.path}><ClinicDetails {...routeProps} {...props} /></Gate>)} />
           <Route path='/clinic-invite' render={routeProps => (<Gate onEnter={boundRequireAuth} key={routeProps.match.path}><ClinicInvite {...routeProps} {...props} /></Gate>)} />
           <Route path='/clinic-workspace/:tab?' render={routeProps => (<Gate onEnter={boundRequireAuth} key={routeProps.match.path}><ClinicWorkspace {...routeProps} {...props} /></Gate>)} />
